requirements.txt
requests
notion-client
langchain
langchain-openai
streamlit
crewai
python-dotenv